+++ docs/ReleaseNotes.md
@@ -15,7 +15,8 @@ The Guava dependency version has been updated to 31.1. Projects may need to chec
 // begin next release
 ### NEXT_RELEASE
 
-* **Feature** Support Greatest and Least scalar function [(Issue #2206)](https://github.com/FoundationDB/fdb-record-layer/issues/2206)
+* **Feature** Support Greatest and Least scalar functions [(Issue #2206)](https://github.com/FoundationDB/fdb-record-layer/issues/2206)
+* **Feature** Support coalesce scalar function [(Issue #2206)](https://github.com/FoundationDB/fdb-record-layer/issues/2207)
 
 * **Bug fix** Fix 1 [(Issue #NNN)](https://github.com/FoundationDB/fdb-record-layer/issues/NNN)
 * **Bug fix** Fix 2 [(Issue #NNN)](https://github.com/FoundationDB/fdb-record-layer/issues/NNN)